Restructure 'GetResourcesFromIdentifiers' query result
Fixes #13

On my Omeka-S 3.1.0 installation (from the zip), it appears fetchAllAssociative
produces the results as an array of 'key->value' pairs representing each
name and value from the 'select' clause of the query, while the code
expects an associative 'identifier -> id' array.

This change makes it work for me again, but of course we should probably
first explain why this didn't seem to have lead to problems for other users
yet.