Match full path when evaluating routes
Previously, with the default configuration of having the default path
for items as:

    document/{item_identifier}

... and the default path for media as:

    document/{item_identifier}/{media_id}

Opening a media page would fail for me: it would correctly generate an
url like `http://mydomain/document/E10/11`, but then the matcher would
already match on the 'item' route and subsequently produce a 'Page not
found', rather than continuing to the 'media' route.

This changes the matching to always match the full path. Since no '/'
characters are allowed in identifiers, the 'item' route regex no longer
matches, and the media route correctly serves that page.