Use 'top' URL when no site-slug is passed in
When rendering the media URL on an Item page like /document/E10,
it seems the 'site-slug' field of 'params' is 'false' when there is only
one site (the default), while it is filled with the slug of the default
site when any additional sites exist. I have not verified whether that
is a bug or intended behavior.

This change makes sure that in that case, the 'top' URL structure is
preferred over the 'public' URL structure, fixing the url to link to
'/document/E10/2' rather than '/s//media/2' (the 'public' URL with the
slug missing).

Fixes #15